docs(plans): the side plan is documentation now, and Platform moves to its home
The second gate round took two findings of eight, and refused the other six --
five of which were one reviewer restating a single objection whose proposed fix
was the code as written, and the sixth would have had the marker compared with
itself before the helper ever ran.

Taken: the Platform union moves out of vendorTerminal, where selection capture
and orphan cleanup had no business depending on the vendor-install module for a
fact about the host. And the plan is promoted, with its deviations recorded --
including the one accepted finding whose stated REASON was measured false, kept
in the record because a plan that adopts a fix while accepting a wrong
explanation for it teaches the next reader the wrong thing.

One tail is owed and named as a tail: a press in a real Remote-WSL window,
which no test can answer.

Observed: plan-lifecycle clean, 1388 passed, 0 failed, 1 skipped.
