+++ research/README.md
@@ -9,6 +9,7 @@ is not finished until it is.
 |---|---|
 | [PLAN_the_tab_wears_an_icon.md](PLAN_the_tab_wears_an_icon.md) | IMPLEMENTED 2026-09-09 — the chat tab wears this product’s own green glyph, baked into a light and a dark file because workbench chrome reaches no theme; the extension URI threaded to the panel, the path in its own module so the resolution is tested rather than the source scanned |
 | [PLAN_a_conversation_survives_a_reload.md](PLAN_a_conversation_survives_a_reload.md) | IMPLEMENTED 2026-09-09 — a window reload keeps every chat tab’s transcript. The conversation carries its own id, because a title-keyed store makes two tabs called `main` overwrite each other before any collision can be detected; nothing is started until the first question, which carries the whole thread across |
+| [PLAN_the_chat_is_correct_on_its_own_side.md](PLAN_the_chat_is_correct_on_its_own_side.md) | IMPLEMENTED 2026-09-10 — `process.platform` describes the extension HOST, not the machine, and four places had been judging the machine by it. `Ctrl+Alt+A` captures from a WSL window through interop (measured at 1.07 s, with `WindowsPowerShell/v1.0` already on the host's own PATH), and a failed capture names which of four things happened rather than blaming the selection; the chat and the `coai-mcp` settings file read THIS side's `vendors`; the orphan sweep ends a matching child on WSL and native Linux, by `comm`/`argv[0]`/`argv[1]` because `codex` and `gemini` are shebang shims. Its two gate rounds refused a pre-flight probe twice and a Blocking finding the plan's own measurement had already refuted |
 | [PLAN_the_gate_diffs_from_a_moving_base.md](PLAN_the_gate_diffs_from_a_moving_base.md) | IMPLEMENTED 2026-09-10 — a code round is a diff against the merge base rather than the tip of a base another session moved, resolved once and used for all three git calls; the fallback pins its ref to a commit, tells a shallow checkout from unrelated histories, and says which it was |
 | [PLAN_a_page_nobody_can_search.md](PLAN_a_page_nobody_can_search.md) | IMPLEMENTED 2026-09-09 — Ctrl+F does something. Every `WebviewPanel` this extension opens is created with `enableFindWidget`, the one option VS Code gives a webview its real find bar for; the guard is a DISCOVERY over every `createWebviewPanel(` call, and its own code round rewrote it after three reviewers showed a string, a comment or a nested object could keep it green over a dead shortcut |
 | [PLAN_the_prompt_box_forgets.md](PLAN_the_prompt_box_forgets.md) | IMPLEMENTED 2026-09-09 — the sidebar prompt box keeps what is typed. The measurement refuted the plan: a chat setting cannot repaint the panel, so the trap it was built around does not exist; the panel gained a third answer beside repaint and patch — WITHHOLD while a control has focus — with the value written as it is typed and the caret returned where it was |
