+++ todo/README.md
@@ -9,7 +9,6 @@ Convention: `.claude/rules/shared/common/planning-docs.md`.
 
 | Plan | What it is for |
 |---|---|
-| [PLAN_the_chat_is_correct_on_its_own_side.md](PLAN_the_chat_is_correct_on_its_own_side.md) | **Bug, three of them with one shape.** `Ctrl+Alt+A` refuses in a WSL window although the editor window is Windows and `powershell.exe` is one interop hop away — measured: the extension host's own PATH carries 33 Windows entries and the helper runs in 1.07 s. The chat and the server settings file read `vendors` off the SHARED configuration, so with *Separate settings for each side* on they launch the CLI of a side this is not — and on this machine `codex` in WSL resolves into the Windows npm directory, so that is not theoretical. And the orphan sweep returns `unknown` on anything but Windows, so a crashed vendor CLI is never ended under WSL. One module names what `process.platform` cannot say, and three callers stop guessing. |
 | [PLAN_the_menu_path_opens_nothing.md](PLAN_the_menu_path_opens_nothing.md) | **Bug, unconfirmed — the code half of the triage is done and clears this product.** Right-click opens the Output console instead of a chat tab; the keybinding opens one. The screenshot's console belongs to a probe extension this repository does not contain, sitting one menu line below ours — This extension creates no output channel at all, and every early return of the menu path raises a notification, so it can produce a tab or a refusal but never a console. What is left is one press with the probe disabled. |
 | [PLAN_a_stale_translation_is_invisible.md](PLAN_a_stale_translation_is_invisible.md) | The help marks a translation that is MISSING and says nothing about one that is BEHIND — `bodyFor` compares keys, never text. Found by the audit of 2026-09-10: five shipped features, four languages describing the tab as it was before any of them, and no note anywhere. A digest of the English body recorded beside each translation, a third outcome on `bodyFor`, and a test that goes red on the commit that makes a translation stale rather than on an audit a day later. |
 | [PLAN_the_log_searches_the_findings.md](PLAN_the_log_searches_the_findings.md) | `findings_fts` has been in the schema since 2026-09-05 and nothing queries it: the search box filters the rows already on screen, so a phrase from a finding's *why* finds nothing. Extracted from the local-db reader plan when it was promoted. |
@@ -42,6 +41,7 @@ in [../research/](../research/) with `IMPLEMENTED` status, beside the record of 
 | 2026-09-10 | [PLAN_provider_then_model.md](../research/PLAN_provider_then_model.md) | Choose the provider, then the model. The list was the configured reviewer ROWS, each labelled with the one model it happened to be set to — so picking another model meant leaving the conversation to reconfigure a reviewer. A provider is a vendor row, settled by three vendors' reviewers overturning the plan; either half of the choice posts both, because the model list belongs to a provider. |
 | 2026-09-10 | [PLAN_who_said_it_and_what_it_cost.md](../research/PLAN_who_said_it_and_what_it_cost.md) | Every answer says which model gave it, in that model's own colour, recorded when the answer arrived rather than looked up later — switching models mid-conversation is a shipped feature, so one tab routinely holds two. And the tab says what the conversation has cost, beside the picker, where the decision to ask again is taken. |
 | 2026-09-10 | [PLAN_a_picture_in_the_question.md](../research/PLAN_a_picture_in_the_question.md) | Paste a screenshot into a question. The mechanism was measured rather than assumed: the file's PATH is named in the prompt and the vendor process opens it, so the file's name, place and lifetime are a contract. `img-src data:` and nothing wider; SVG refused although it is an image; and a provider that cannot take one is refused BY NAME, because a picture that silently does not arrive is the worst outcome this feature has. |
+| 2026-09-10 | [PLAN_the_chat_is_correct_on_its_own_side.md](../research/PLAN_the_chat_is_correct_on_its_own_side.md) | Four defects with one shape: code judging THE MACHINE by `process.platform`, which describes only the extension host. `Ctrl+Alt+A` captures from a WSL window through interop — measured at 1.07 s against a 6 s cap, with 33 Windows PATH entries in the host's own environment — and a failed capture now says which of four things happened instead of blaming the selection every time. The chat and the `coai-mcp` settings file read THIS side's `vendors`, so a WSL window stops launching the Windows npm shim `codex` resolves to there, and the gate stops running reviewers off another side's rows. The orphan sweep ends a matching child on WSL and on native Linux, identifying it by `comm`/`argv[0]`/`argv[1]` because `codex` and `gemini` are `#!/bin/sh` shims whose `argv[0]` is the interpreter. Two gate rounds took 23 findings and refused 20, including a Blocking claim its own opening measurement had already refuted. |
 | 2026-09-10 | [PLAN_the_gate_diffs_from_a_moving_base.md](../research/PLAN_the_gate_diffs_from_a_moving_base.md) | A code round is a diff against the MERGE BASE. `A..B` shows the commits the base has and the branch does not as deletions the branch performed, and three times a round judged another session’s merge as this branch’s work — 1616 phantom deletions against 12 real ones, filed as Blocking by two vendors. Three call sites, not the two the plan counted: `cat-file` takes a rev and has no three-dot form. Its own code round then found the same defect one layer down, in a fallback that kept a moving ref — which closed an argument injection with it. |
 | 2026-09-09 | [PLAN_an_answer_reads_like_a_document.md](../research/PLAN_an_answer_reads_like_a_document.md) | A model's answer is rendered rather than shown as its own source: headings, nested lists, code in its own scrolling box, tables, links that open in the browser or in the editor at the line they name. `marked` tokenizes and the emitting is ours, against a fixed list of tags — the plan round refused the zero-dependency alternative, because *escape first, mark up second* as a regex pipeline corrupts a fence and cannot count nesting. No `href` is ever emitted; the host resolves every link and decides. The hostile-input file found GFM autolinking and a weakness in its own assertions on its first run. |
 | 2026-09-09 | [PLAN_the_tab_wears_an_icon.md](../research/PLAN_the_tab_wears_an_icon.md) | The chat tab wears this product’s own glyph, in a light and a dark file, because tab chrome cannot read a theme. The `iconPath` line was the small half: `createChatPanel` had no extension URI and neither had either of its callers. The path lives in its own module so the RESOLUTION can be tested rather than the source scanned — and the packaging guard that says the files ship was itself green over `**/*.svg` until the code round caught it. |
